Implement Release Pipeline v4 maximum-speed delivery
