+++ .github/actions-security-policy.json
@@ -19,9 +19,6 @@
       "actions",
       "contents"
     ],
-    ".github/workflows/build-release-pipeline-v4-candidate-v3.yml": [
-      "contents"
-    ],
     ".github/workflows/discord-release-preview.yml": [
       "issues"
     ],
