--- .github/workflows/diagnose-release-pipeline-v4-contract-v2.yml
@@ -1,74 +1,0 @@
-name: Diagnose Pipeline v4 Contract Reconciliation
-
-on:
-  pull_request:
-    branches: [main]
-    paths:
-      - ".github/workflows/diagnose-release-pipeline-v4-contract-v2.yml"
-      - ".github/development-packages/issue543_release_pipeline_v4.py"
-
-permissions:
-  contents: read
-
-jobs:
-  diagnose:
-    if: github.head_ref == 'feature/release-pipeline-v4-fast-lane'
-    runs-on: ubuntu-latest
-    timeout-minutes: 10
-    steps:
-      - uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7
-        with:
-          fetch-depth: 0
-          persist-credentials: false
-
-      - name: Run exact reconciliation and retain output
-        shell: bash
-        run: |
-          set +e
-          {
-            python3 - <<'PY'
-          from pathlib import Path
-          path = Path('.github/development-packages/issue543_release_pipeline_v4.py')
-          text = path.read_text(encoding='utf-8')
-          fixes = (
-              ('path.write_text(json.dumps(data,indent=2)+"\\n",encoding="utf-8")',
-               'path.write_text(json.dumps(data,indent=2)+"\\\\n",encoding="utf-8")'),
-              ('(root/"status/RELEASE_SPEED.md").write_text("\\n".join(lines),encoding="utf-8")',
-               '(root/"status/RELEASE_SPEED.md").write_text("\\\\n".join(lines),encoding="utf-8")'),
-          )
-          for old, new in fixes:
-              print('escape', repr(old), text.count(old))
-              if text.count(old) != 1: raise SystemExit('escape mismatch')
-              text = text.replace(old, new, 1)
-          path.write_text(text, encoding='utf-8')
-          PY
-            python3 .github/development-packages/issue543_release_pipeline_v4.py
-            python3 - <<'PY'
-          from pathlib import Path
-          path = Path('.github/scripts/test_version_status_contract.py')
-          text = path.read_text(encoding='utf-8')
-          checks = (
-              ('- name: Record successful release, manifest and announcement state', 2),
-              ('- name: Publish GitHub Pages', 2),
-              ('        \'gh run watch "$PAGES_RUN_ID" --exit-status\',\n', 1),
-              ('        "PAGES_RUN_ID: ${{ steps.pages.outputs.pages_run_id }}",\n', 1),
-              ('        "manifest_run_id",\n', 1),
-          )
-          for token, expected in checks:
-              print('token', repr(token), 'count', text.count(token), 'expected', expected)
-          PY
-          } > pipeline-v4-contract-v2.log 2>&1
-          echo $? > pipeline-v4-contract-v2-exit.txt
-          cat pipeline-v4-contract-v2.log
-          exit 0
-
-      - name: Upload reconciliation diagnostics
-        if: always()
-        uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
-        with:
-          name: release-pipeline-v4-contract-v2-${{ github.sha }}
-          path: |
-            pipeline-v4-contract-v2.log
-            pipeline-v4-contract-v2-exit.txt
-          if-no-files-found: error
-          retention-days: 14
