--- .github/workflows/diagnose-release-pipeline-v4-final.yml
@@ -1,107 +1,0 @@
-name: Diagnose Final Pipeline v4 Correction
-
-on:
-  pull_request:
-    branches: [main]
-    paths:
-      - ".github/workflows/diagnose-release-pipeline-v4-final.yml"
-      - ".github/development-packages/issue543_release_pipeline_v4.py"
-
-permissions:
-  contents: read
-
-jobs:
-  diagnose:
-    if: github.head_ref == 'feature/release-pipeline-v4-fast-lane'
-    runs-on: ubuntu-latest
-    timeout-minutes: 10
-    steps:
-      - uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7
-        with:
-          fetch-depth: 0
-          persist-credentials: false
-
-      - name: Run exact final correction
-        shell: bash
-        run: |
-          set +e
-          {
-            python3 - <<'PY'
-          from pathlib import Path
-          path = Path('.github/development-packages/issue543_release_pipeline_v4.py')
-          text = path.read_text(encoding='utf-8')
-          fixes = (
-              ('path.write_text(json.dumps(data,indent=2)+"\\n",encoding="utf-8")', 'path.write_text(json.dumps(data,indent=2)+"\\\\n",encoding="utf-8")'),
-              ('(root/"status/RELEASE_SPEED.md").write_text("\\n".join(lines),encoding="utf-8")', '(root/"status/RELEASE_SPEED.md").write_text("\\\\n".join(lines),encoding="utf-8")'),
-          )
-          for old, new in fixes:
-              print('package token', repr(old), text.count(old))
-              if text.count(old) != 1: raise SystemExit('package token mismatch')
-              text = text.replace(old, new, 1)
-          path.write_text(text, encoding='utf-8')
-          PY
-            python3 .github/development-packages/issue543_release_pipeline_v4.py
-            python3 - <<'PY'
-          from pathlib import Path
-          path = Path('.github/scripts/test_version_status_contract.py')
-          text = path.read_text(encoding='utf-8')
-          replacements = (
-              ('- name: Record successful release, manifest and announcement state', '- name: Record successful release, manifest, announcement and speed state', 2),
-              ('- name: Publish GitHub Pages', '- name: Dispatch GitHub Pages asynchronously', 2),
-              ('        \'gh run watch "$PAGES_RUN_ID" --exit-status\',\n', '', 1),
-          )
-          for old, new, expected in replacements:
-              print('contract token', repr(old), text.count(old), expected)
-              if text.count(old) != expected: raise SystemExit('contract token mismatch')
-              text = text.replace(old, new)
-          lines = text.splitlines(keepends=True)
-          indexes = [i for i, line in enumerate(lines) if '"PAGES_RUN_ID:' in line]
-          print('pages indexes', indexes)
-          if len(indexes) != 1: raise SystemExit('pages index mismatch')
-          expression = '$' + '{{ steps.pages.outputs.dispatched }}'
-          lines[indexes[0]] = '        "PAGES_DISPATCHED: ' + expression + '",\n'
-          text = ''.join(lines)
-          anchor = '        "manifest_run_id",\n'
-          print('anchor', text.count(anchor))
-          if text.count(anchor) != 1: raise SystemExit('anchor mismatch')
-          text = text.replace(anchor, anchor + '        \'gh run watch "$PAGES_RUN_ID" --exit-status\',\n', 1)
-          path.write_text(text, encoding='utf-8')
-
-          release_path = Path('.github/workflows/release-toolkit.yml')
-          release = release_path.read_text(encoding='utf-8')
-          print('control chars', release.count(chr(1)))
-          if release.count(chr(1)) != 1: raise SystemExit('control char mismatch')
-          release = release.replace(chr(1), r'\1', 1)
-          old = '        env:\n          RELEASE_VERSION: ${{ inputs.version }}\n          BACKUP_COMMIT: ${{ steps.external.outputs.backup_commit }}\n        env:\n          RELEASE_STARTED_EPOCH: ${{ steps.release_start.outputs.started_epoch }}\n'
-          new = '        env:\n          RELEASE_VERSION: ${{ inputs.version }}\n          BACKUP_COMMIT: ${{ steps.external.outputs.backup_commit }}\n          RELEASE_STARTED_EPOCH: ${{ steps.release_start.outputs.started_epoch }}\n'
-          print('duplicate env', release.count(old))
-          if release.count(old) != 1: raise SystemExit('duplicate env mismatch')
-          release = release.replace(old, new, 1)
-          old = '            echo "- ✅ Greasy Fork verification and private backup ran concurrently\n            echo "- ✅ Greasy Fork version verified""\n'
-          new = '            echo "- ✅ Greasy Fork verification and private backup ran concurrently"\n            echo "- ✅ Greasy Fork version verified"\n'
-          print('summary', release.count(old))
-          if release.count(old) != 1: raise SystemExit('summary mismatch')
-          release = release.replace(old, new, 1)
-          old = '            echo "- ✅ GitHub Pages deployment completed: ${PAGES_RUN_ID}"\n'
-          new = '            echo "- ✅ GitHub Pages deployment dispatched asynchronously: ${PAGES_DISPATCHED}"\n'
-          print('pages summary', release.count(old))
-          if release.count(old) != 1: raise SystemExit('pages summary mismatch')
-          release = release.replace(old, new, 1)
-          release_path.write_text(release, encoding='utf-8')
-          print('correction passed')
-          PY
-          } > pipeline-v4-final-correction.log 2>&1
-          echo $? > pipeline-v4-final-correction-exit.txt
-          cat pipeline-v4-final-correction.log
-          exit 0
-
-      - name: Upload exact correction diagnostics
-        if: always()
-        uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
-        with:
-          name: release-pipeline-v4-final-correction-${{ github.sha }}
-          path: |
-            pipeline-v4-final-correction.log
-            pipeline-v4-final-correction-exit.txt
-          if-no-files-found: error
-          retention-days: 14
