--- .github/workflows/diagnose-release-pipeline-v4-materialise.yml
@@ -1,71 +1,0 @@
-name: Diagnose Pipeline v4 Materialisation
-
-on:
-  pull_request:
-    branches: [main]
-    paths:
-      - ".github/workflows/diagnose-release-pipeline-v4-materialise.yml"
-      - ".github/development-packages/issue543_release_pipeline_v4.py"
-
-permissions:
-  contents: read
-
-jobs:
-  diagnose:
-    if: github.head_ref == 'feature/release-pipeline-v4-fast-lane'
-    runs-on: ubuntu-latest
-    timeout-minutes: 10
-    steps:
-      - uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7
-        with:
-          fetch-depth: 0
-          persist-credentials: false
-
-      - name: Materialise and capture output
-        shell: bash
-        run: |
-          set +e
-          {
-            python3 - <<'PY'
-          from pathlib import Path
-          path = Path('.github/development-packages/issue543_release_pipeline_v4.py')
-          text = path.read_text(encoding='utf-8')
-          fixes = (
-              ('path.write_text(json.dumps(data,indent=2)+"\\n",encoding="utf-8")',
-               'path.write_text(json.dumps(data,indent=2)+"\\\\n",encoding="utf-8")'),
-              ('(root/"status/RELEASE_SPEED.md").write_text("\\n".join(lines),encoding="utf-8")',
-               '(root/"status/RELEASE_SPEED.md").write_text("\\\\n".join(lines),encoding="utf-8")'),
-          )
-          for old, new in fixes:
-              if text.count(old) != 1:
-                  raise SystemExit(f'Unexpected embedded escape token count: {old!r} -> {text.count(old)}')
-              text = text.replace(old, new, 1)
-          path.write_text(text, encoding='utf-8')
-          PY
-            python3 .github/development-packages/issue543_release_pipeline_v4.py
-            python3 - <<'PY'
-          from pathlib import Path
-          path = Path('.github/scripts/test_version_status_contract.py')
-          text = path.read_text(encoding='utf-8')
-          old = '- name: Record successful release, manifest and announcement state'
-          new = '- name: Record successful release, manifest, announcement and speed state'
-          print('marker count', text.count(old))
-          if text.count(old) != 1:
-              raise SystemExit(f'Unexpected version-status marker count: {text.count(old)}')
-          path.write_text(text.replace(old, new, 1), encoding='utf-8')
-          PY
-          } > pipeline-v4-materialise.log 2>&1
-          echo $? > pipeline-v4-materialise-exit.txt
-          cat pipeline-v4-materialise.log
-          exit 0
-
-      - name: Upload materialisation diagnostics
-        if: always()
-        uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
-        with:
-          name: release-pipeline-v4-materialise-${{ github.sha }}
-          path: |
-            pipeline-v4-materialise.log
-            pipeline-v4-materialise-exit.txt
-          if-no-files-found: error
-          retention-days: 14
