--- .github/workflows/preflight-release-pipeline-v4.yml
@@ -1,104 +1,0 @@
-name: Preflight Release Pipeline v4
-
-on:
-  pull_request:
-    branches: [main]
-    paths:
-      - ".github/development-packages/issue543_release_pipeline_v4.py"
-      - ".github/workflows/preflight-release-pipeline-v4.yml"
-
-permissions:
-  contents: read
-
-jobs:
-  preflight:
-    if: github.head_ref == 'feature/release-pipeline-v4-fast-lane'
-    runs-on: ubuntu-latest
-    timeout-minutes: 15
-    steps:
-      - name: Check out exact candidate
-        uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7
-        with:
-          fetch-depth: 0
-          persist-credentials: false
-
-      - name: Apply reviewed Pipeline v4 package locally
-        id: apply
-        shell: bash
-        run: |
-          set +e
-          python3 - <<'PY'
-          from pathlib import Path
-          path = Path('.github/development-packages/issue543_release_pipeline_v4.py')
-          text = path.read_text(encoding='utf-8')
-          fixes = (
-              ('path.write_text(json.dumps(data,indent=2)+"\\n",encoding="utf-8")',
-               'path.write_text(json.dumps(data,indent=2)+"\\\\n",encoding="utf-8")'),
-              ('(root/"status/RELEASE_SPEED.md").write_text("\\n".join(lines),encoding="utf-8")',
-               '(root/"status/RELEASE_SPEED.md").write_text("\\\\n".join(lines),encoding="utf-8")'),
-          )
-          for old, new in fixes:
-              if text.count(old) != 1:
-                  raise SystemExit(f'Unexpected embedded escape token count: {old!r} -> {text.count(old)}')
-              text = text.replace(old, new, 1)
-          path.write_text(text, encoding='utf-8')
-          PY
-          python3 .github/development-packages/issue543_release_pipeline_v4.py > pipeline-v4-apply.log 2>&1
-          status=$?
-          cat pipeline-v4-apply.log
-          echo "exit_code=$status" >> "$GITHUB_OUTPUT"
-          if [[ "$status" -eq 0 ]]; then
-            rm -f .github/development-packages/issue543_release_pipeline_v4.py
-            rm -f .github/workflows/preflight-release-pipeline-v4.yml
-            find .github/development-packages -type d -empty -delete 2>/dev/null || true
-          fi
-          exit 0
-
-      - name: Upload package diagnostics
-        if: always()
-        uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
-        with:
-          name: release-pipeline-v4-package-diagnostics-${{ github.sha }}
-          path: pipeline-v4-apply.log
-          if-no-files-found: error
-          retention-days: 14
-
-      - name: Stop on package failure
-        if: steps.apply.outputs.exit_code != '0'
-        run: exit 1
-
-      - name: Validate Pipeline v4 and retained Toolkit contracts
-        run: |
-          set -euo pipefail
-          python3 .github/scripts/test_release_pipeline_v4.py
-          python3 .github/scripts/validate_userscript.py
-          node --check src/MissionChief_Map_Command_Toolkit.user.js
-          cmp --silent dist/MissionChief_Map_Command_Toolkit.user.js dist/MissionChief_Map_Command_Toolkit.txt
-
-      - name: Build exact file manifest
-        run: |
-          set -euo pipefail
-          mkdir -p pipeline-v4-candidate
-          for file in \
-            .github/workflows/validate-userscript.yml \
-            .github/workflows/auto-release-after-validation.yml \
-            .github/workflows/release-toolkit.yml \
-            .github/scripts/run_userscript_preflight.sh \
-            .github/scripts/record_release_speed.py \
-            .github/scripts/generate_release_speed_dashboard.py \
-            .github/scripts/test_release_pipeline_v4.py \
-            status/release-speed-history.json \
-            status/RELEASE_SPEED.md \
-            docs/RELEASE_PIPELINE_V4.md; do
-            mkdir -p "pipeline-v4-candidate/$(dirname "$file")"
-            cp "$file" "pipeline-v4-candidate/$file"
-          done
-          find pipeline-v4-candidate -type f -print0 | sort -z | xargs -0 sha256sum > pipeline-v4-candidate/SHA256SUMS.txt
-
-      - name: Upload exact validated Pipeline v4 candidate
-        uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
-        with:
-          name: release-pipeline-v4-candidate-${{ github.sha }}
-          path: pipeline-v4-candidate/
-          if-no-files-found: error
-          retention-days: 14
